// gtfs-processor/shared/gtfs-db-types/types.ts
export enum CongestionLevel {
  UNKNOWN_CONGESTION_LEVEL = "UNKNOWN_CONGESTION_LEVEL",
  RUNNING_SMOOTHLY = "RUNNING_SMOOTHLY",
  STOP_AND_GO = "STOP_AND_GO",
  CONGESTION = "CONGESTION",
  SEVERE_CONGESTION = "SEVERE_CONGESTION",
}

export enum Direction {
  OUTBOUND = "OUTBOUND",
  INBOUND = "INBOUND",
}

export enum LocationType {
  STOP_OR_PLATFORM = "STOP_OR_PLATFORM",
  STATION = "STATION",
  ENTRANCE_EXIT = "ENTRANCE_EXIT",
  GENERIC_NODE = "GENERIC_NODE",
  BOARDING_AREA = "BOARDING_AREA",
}

export enum OccupancyStatus {
  EMPTY = "EMPTY",
  MANY_SEATS_AVAILABLE = "MANY_SEATS_AVAILABLE",
  FEW_SEATS_AVAILABLE = "FEW_SEATS_AVAILABLE",
  STANDING_ROOM_ONLY = "STANDING_ROOM_ONLY",
  CRUSHED_STANDING_ROOM_ONLY = "CRUSHED_STANDING_ROOM_ONLY",
  FULL = "FULL",
  NOT_ACCEPTING_PASSENGERS = "NOT_ACCEPTING_PASSENGERS",
  NO_DATA_AVAILABLE = "NO_DATA_AVAILABLE",
  NOT_BOARDABLE = "NOT_BOARDABLE",
}
// ...
export enum VehicleStopStatus {
  INCOMING_AT = "INCOMING_AT",
  STOPPED_AT = "STOPPED_AT",
  IN_TRANSIT_TO = "IN_TRANSIT_TO",
}
// ...
export interface VehiclePosition {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  vehicle_id: string;
  timestamp?: Date;
  stop_id: null | string;
  current_stop_sequence: null | number;
  current_status: VehicleStopStatus | null;
  congestion_level: CongestionLevel | null;
  occupancy_status: OccupancyStatus | null;
  occupancy_percentage: null | number;
  latitude: null | number;
  longitude: null | number;
  bearing: null | number;
  odometer: null | number;
  speed: null | number;
  ingested_at?: Date;
  trip?: any | null;
}